created mutations to manager available in ad
