+++ src/graphql/resolvers/Mutations/Mutation.ts
@@ -1,0 +1,16 @@
+import { MutationResolvers } from '../../../generated/resolvers/resolverTypes';
+import createBooking from './createBooking';
+import signIn from './signIn';
+import signUp from './signUp';
+import updateBlockedDay from './updateBlockedDay';
+import deleteBlockedDay from './deleteBlockedDay';
+
+const Mutation: MutationResolvers.Type = {
+  signIn,
+  signUp,
+  createBooking,
+  updateBlockedDay,
+  deleteBlockedDay,
+};
+
+export default Mutation;
