+++ src/graphql/resolvers/Mutations/createBooking.ts
@@ -1,0 +1,50 @@
+import moment from 'moment';
+import { MutationResolvers } from '../../../generated/resolvers/resolverTypes';
+import CustomError from '../../../model/error/CustomError';
+import { Context } from '../../../context';
+import { Booking } from '../../../generated/prisma-client';
+
+const createBooking = async (
+  _: undefined,
+  {
+    data: { adId, checkin, checkout, clientId },
+  }: MutationResolvers.ArgsCreateBooking,
+  { prisma }: Context,
+): Promise<Booking> => {
+  const overlapDays = await prisma.ad({ id: adId }).blockedDays({
+    where: {
+      NOT: [
+        {
+          OR: [
+            {
+              endDay_lt: checkin,
+            },
+            { startDay_gt: checkout },
+          ],
+        },
+      ],
+    },
+  });
+  if (overlapDays.length > 0) throw new CustomError('OVERLAP_BOOKING');
+  const price = (await prisma.ad({ id: adId }))?.price;
+  if (!price) throw new CustomError('AD_NOT_FOUND');
+  const nights = moment(checkout).diff(checkin, 'days');
+  const totalPaid = nights * price;
+  const booking = await prisma.createBooking({
+    checkin,
+    checkout,
+    totalPaid,
+    pax: 2,
+    client: { connect: { id: clientId } },
+    ad: { connect: { id: adId } },
+  });
+  await prisma.createBlockedDay({
+    startDay: checkin,
+    endDay: checkout,
+    byBooking: true,
+    ad: { connect: { id: adId } },
+  });
+  return booking;
+};
+
+export default createBooking;
